[refactor/#304] 마이페이지 비로그인 리다이렉트 서버로 이동, route 변경 (빌드 경고 해결)
