--- src/app/user/me/layout.tsx
@@ -1,8 +1,0 @@
-'use client'
-
-import DefaultQueryProvider from '@/lib/queries/DefaultQueryProvider'
-import { ChildrenProps } from '@/lib/types/types'
-
-export default function FormLayout({ children }: ChildrenProps) {
-  return <DefaultQueryProvider>{children}</DefaultQueryProvider>
-}
